feature: Support multiple authorizers
This adds a new authorize trait to simply authorize based on a token and
refactors the layer to wrap anything implementing this trait. The trait
is ofcourse implemented for Authorizer (equivalent to the current code)
as well as slices and vectors of Authorizers (or really anything
implement authorize) enabling the use of multiple authorities.

To support creating authorizers the main builder is split up into a
builder for authorize (to create one or more) and a seperate small one
for the layer itself just for the jwt source.

All Authorizers used should be setup for the same claims so a consistent
TokenData can be passed down.

Fixes: #14

Signed-off-by: Sjoerd Simons <sjoerd@collabora.com>